переделка

+++ src/main/java/cosmonaut/service/OrderService.java
@@ -1,15 +1,15 @@
 package cosmonaut.service;
 
-import cosmonaut.entity.Authority;
-import cosmonaut.entity.Order;
-import cosmonaut.entity.OrderItem;
-import cosmonaut.entity.User;
+import cosmonaut.entity.*;
+import cosmonaut.entity.enums.UserRole;
 import cosmonaut.repository.OrderRepository;
+import cosmonaut.util.CurrentUserUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.security.Principal;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -18,6 +18,8 @@ public class OrderService {
     private OrderRepository orderRepository;
 
     private UserService userService;
+
+    private CurrentUserUtils currentUserUtils;
 
     @Autowired
     public void setUserService(UserService userService) {
@@ -29,15 +31,20 @@ public class OrderService {
         this.orderRepository = orderRepository;
     }
 
+    @Autowired
+    public void setCurrentUserUtils(CurrentUserUtils currentUserUtils) {
+        this.currentUserUtils = currentUserUtils;
+    }
+
     public Order createOrderFromItems(User user, List<OrderItem> orderItems) {
         Order order = new Order();
         order.setOrderItems(new ArrayList<>());
         order.setUser(user);
-        orderItems.stream().forEach(orderItem -> {
+        order.setIsAccepted(false);
+        orderItems.forEach(orderItem -> {
             order.getOrderItems().add(orderItem);
             orderItem.setOrder(order);
         });
-        orderItems.clear();
         return orderRepository.save(order);
     }
 
@@ -53,20 +60,29 @@ public class OrderService {
         return orderRepository.findAll();
     }
 
-    public void deleteOrderById(Long id) {
-        orderRepository.deleteById(id);
+    public void acceptOrderById(Long id) {
+        Order orderToAccept = orderRepository.findById(id).get();
+        UserRole role = currentUserUtils.getCurrentLoggedUser().getRole();
+        if (role == UserRole.MANAGER) {
+            orderToAccept.setIsAccepted(true);
+            orderRepository.save(orderToAccept);
+        }
     }
 
-    public List<Order> getCustomOrders(Principal principal) {
-        User user = userService.findByUsername(principal.getName());
-        List<Authority> authorities = user.getAuthorities();
-        for (Authority authority : authorities) {
-            if (authority.getAuthority().equals("ROLE_SELLER")) {
-                return getAllOrders();
-            }
+    public void deleteOrderById(Long id) {
+        Order orderToDelete = orderRepository.findById(id).get();
+        if(!orderToDelete.getIsAccepted()) {
+            orderRepository.deleteById(id);
+        }
+    }
+
+    public List<Order> getCustomOrders() {
+        User user = currentUserUtils.getCurrentLoggedUser();
+        if(user == null) return Collections.emptyList();
+        UserRole role = user.getRole();
+        if (role == UserRole.ADMIN || role == UserRole.MANAGER) {
+            return getAllOrders();
         }
         return getOrderByUser(user);
     }
-
-
 }
