+++ src/main/java/cosmonaut/service/ProductService.java
@@ -1,20 +1,30 @@
 package cosmonaut.service;
 
 import cosmonaut.entity.Product;
+import cosmonaut.entity.User;
+import cosmonaut.entity.enums.UserRole;
 import cosmonaut.repository.ProductRepository;
+import cosmonaut.util.CurrentUserUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+
+import java.util.Collections;
 
 @Service
 public class ProductService {
 
     private ProductRepository productRepository;
 
+    private CurrentUserUtils currentUserUtils;
+
     @Autowired
-    public void setProductRepository(ProductRepository productRepository) {
-        this.productRepository = productRepository;
+    public void setProductRepository(ProductRepository productRepository) { this.productRepository = productRepository; }
+
+    @Autowired
+    public void setCurrentUserUtils(CurrentUserUtils currentUserUtils) {
+        this.currentUserUtils = currentUserUtils;
     }
 
     public Page<Product> getAllProducts(Pageable pageable) {
@@ -28,16 +38,12 @@ public class ProductService {
         return null;
     }
 
-    public Product getProductByTitle(String title) {
-        return productRepository.findOneByTitle(title);
-    }
-
     public void deleteProductById(Long id) {
-        productRepository.deleteById(id);
-    }
-
-    // In future releases
-    public void addProduct(Product product) {
-        productRepository.save(product);
+        User user = currentUserUtils.getCurrentLoggedUser();
+        if(user == null) return;
+        UserRole role = user.getRole();
+        if (role == UserRole.ADMIN) {
+            productRepository.deleteById(id);
+        }
     }
 }
